+++ common/scala/src/main/resources/reference.conf
@@ -7,4 +7,6 @@ whisk.spi {
   ContainerFactoryProvider = whisk.core.containerpool.docker.DockerContainerFactoryProvider
   LogStoreProvider = whisk.core.containerpool.logging.DockerToActivationLogStoreProvider
   LoadBalancerProvider = whisk.core.loadBalancer.ShardingContainerPoolBalancer
+  ContainerPoolProvider = whisk.core.containerpool.DefaultContainerPool
+  ContainerProxyProvider = whisk.core.containerpool.DefaultContainerProxy
 }
